iorap: Add iorap.inode2filename command line binary
Maps a list of inodes into their filenames by scanning the filesystem.

This is a separate binary that will be fork+execd by iorapd as a background
job when it occasionally needs to map ftrace events (from perfetto)
from inode numbers into actual filenames.

The inode scanning is following a security requirement to be in a separate
process, and an sepolicy CL will be added in a follow-up to grant limit
file scanning privileges to this binary.

Test: atest iorapd-tests # and also manual
Bug: 72170747

Change-Id: I280aa62ce3cc1d7efdadc22b40bd1435496b7086