compiler: Add --inode-textcache flag [deterministic compilation]
Equivalent to inode2filename --verify=none --textcache=<xyz>.
This makes compilation deterministic.

When not using this flag, it defaults to doing a diskscan which
can cause the compilation output to change from run to run as inodes
on the filesystem get modified.

Also speeds up compilation slightly since looking up from a cache
is faster than scanning the whole disk.

Bug: 72170747
Test: adb shell iorap.cmd.compiler /data/misc/iorapd/com.google.android.dialer%2Fcom.google.android.dialer.extensions.GoogleDialtactsActivity.perfetto_trace.pb -it /data/local/tmp/dumpcache -op /data/local/tmp/compiler_out2.pb
Change-Id: I253d60acd6fb3eec8b6715c1d1882eb606ba67cb